Translated buttons response
