+++ bot/locale/en_US.json
@@ -53,6 +53,9 @@
   "LOGIN_RESPONSE": "Go to the website and authenticate",
   "LOGOUT_RESPONSE": "Are you sure you want to log out of the `%nickname%` penguin?",
   "ONLY_ONE_PENGUIN_LINKED": "You only have one penguin linked. You can link more penguins using command </login:1099629339110289442>.",
+  "LOGOUT_SUCCESS": "Your penguin `%nickname%` has been successfully unlinked",
+  "LOGOUT_SUCCESS_ALT": "Your penguin `%nickname%` has been successfully unlinked. Your current penguin is now `%new_nickname%`",
+  "LOGOUT_SUCCESS_ALT2": "Your penguin `%nickname%` has been successfully unlinked. To select your current penguin, use </switch:1100480798647398422>",
   "SWITCH_RESPONSE": "Your current penguin: `%nickname%`. Which penguin do you want to make the current penguin?",
   "SWITCH_RESPONSE_ALT": "Your current penguin is not selected. Which penguin do you want to make the current penguin?",
   "SETTINGS_RESPONSE": "You can change your settings here. \n# Notifications",
