+++ bot/locale/ru.json
@@ -54,6 +54,9 @@
   "LOGIN_RESPONSE": "Перейдите на сайт и пройдите авторизацию",
   "LOGOUT_RESPONSE": "Вы уверены, что хотите выйти с пингвина `%nickname%`?",
   "ONLY_ONE_PENGUIN_LINKED": "У вас привязан только один пингвин. Вы можете привязать ещё несколько с помощью команды </login:1099629339110289442>",
+  "LOGOUT_SUCCESS": "Ваш пингвин `%nickname%` успешно отвязан",
+  "LOGOUT_SUCCESS_ALT": "Ваш пингвин `%nickname%` успешно отвязан. Сейчас ваш текущего пингвина `%new_nickname%`",
+  "LOGOUT_SUCCESS_ALT2": "Ваш пингвин `%nickname%` успешно отвязан. Чтобы выбрать текущего пингвина, воспользуйтесь командой </switch:1100480798647398422>",
   "SWITCH_RESPONSE": "Ваш текущий пингвин: `%nickname%`. Какого пингвина вы хотите сделать текущим?",
   "SWITCH_RESPONSE_ALT": "Ваш текущий пингвин не выбран. Какого пингвина вы хотите сделать текущим?",
   "SETTINGS_RESPONSE": "Можете изменить свои настройки здесь.\n# Уведомления",
